Deduplicate the user principal authorization gate in repo-store kinds
The four repo-store kind handlers each carried a copy of the bearer
token claims and pre-resolved authz cross-check. They now delegate to
one helper parameterized by the verdict resource prefix, so the
security-critical gate cannot drift between kinds. Behavior, denial
strings, and exported signatures are unchanged.
