+++ packages/hub-sessions/src/repo-store/index.ts
@@ -19,6 +19,10 @@ export type {
   WriteTreePreservingPrefixArgs,
 } from "./types";
 export { UserPrincipal } from "./types";
+export {
+  authorizeUserPrincipal,
+  type AuthorizeUserPrincipalArgs,
+} from "./user-principal-gate";
 export { createRepoStore, type CreateRepoStoreConfig } from "./store";
 export {
   subscribeKind,
